aider: Moved import statements to the top of the file to fix syntax error.
# Aider chat conversation:

USER: curseforge/curseforge.go:178:1: syntax error: imports must appear before other declarations

ASSISTANT: To fix the syntax error, we need to move the import statement to the top of the file. Here are the necessary *SEARCH/REPLACE* blocks:

First, we will remove the misplaced import statement:

curseforge/curseforge.go
```go
<<<<<<< SEARCH
import (
	"encoding/json"
	"os"
)

=======
>>>>>>> REPLACE
```

Next, we will add the import statement to the top of the file:

curseforge/curseforge.go
```go
<<<<<<< SEARCH
	"github.com/mitchellh/mapstructure"
	"github.com/packwiz/packwiz/cmd"
	"github.com/packwiz/packwiz/core"
	"github.com/spf13/cobra"
)

=======
	"encoding/json"
	"os"

	"github.com/mitchellh/mapstructure"
	"github.com/packwiz/packwiz/cmd"
	"github.com/packwiz/packwiz/core"
	"github.com/spf13/cobra"
>>>>>>> REPLACE
```

This will ensure that the import statements are correctly placed at the top of the file.